fix re-structure app
