+++ src/assets/icon_pin.png
