+++ src/assets/icon_restauration.png
