+++ src/assets/icon_supply.png
