+++ src/screens/CardScreenInfo.js
