+++ src/services/shop-service.ts
