+++ src/store/enums/shop-activity.enum.ts
