+++ src/store/interfaces/card.interface.ts
