+++ src/store/interfaces/index.ts
