+++ src/store/interfaces/promotion.interface.ts
