+++ src/store/interfaces/shop.interface.ts
