+++ src/store/interfaces/utility.interface.ts
