+++ src/store/reducers/index.ts
