--- store/reducers/username.js
@@ -1,7 +1,0 @@
-export default function (username = "", action) {
-  if (action.type === "saveUsername") {
-    return action.username;
-  } else {
-    return username;
-  }
-}
